backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
firebase-admin==6.4.0
google-cloud-aiplatform==1.38.1
vertexai==1.71.1
reportlab==4.0.6
# Pillow==10.0.1
google-cloud-videointelligence==2.8.0
google-cloud-vision==3.4.0
google-generativeai==0.3.2
protobuf==4.25.3